+++ android/expoview/src/main/java/versioned/host/exp/exponent/modules/api/screens/ScreenStackFragment.java
@@ -109,10 +109,11 @@ public class ScreenStackFragment extends ScreenFragment {
     }
   }
 
-  public void onStackUpdate() {
-    View child = mScreenView.getChildAt(0);
-    if (child instanceof ScreenStackHeaderConfig) {
-      ((ScreenStackHeaderConfig) child).onUpdate();
+  @Override
+  public void onContainerUpdate() {
+    ScreenStackHeaderConfig headerConfig = getScreen().getHeaderConfig();
+    if (headerConfig != null) {
+      headerConfig.onUpdate();
     }
   }