+++ docs/components/plugins/TerminalBlock.tsx
@@ -1,11 +1,11 @@
 import { css } from '@emotion/core';
-import { colors } from '@expo/styleguide';
+import { palette } from '@expo/styleguide';
 import * as React from 'react';
 
 import * as Constants from '~/constants/theme';
 
 const STYLES_PROMPT = css`
-  background-color: ${colors.light.black};
+  background-color: ${palette.light.black};
   border-radius: 4px;
   padding: 24px;
   display: flex;
@@ -18,11 +18,11 @@ const STYLES_LINE = css`
   white-space: nowrap;
   font-family: ${Constants.fontFamilies.mono};
   font-size: 13px;
-  color: ${colors.dark.gray[900]};
+  color: ${palette.dark.gray[900]};
   line-height: 160%;
   ::before {
     content: '$ ';
-    color: ${colors.dark.primary[700]};
+    color: ${palette.dark.primary[700]};
   }
 `;
 
@@ -31,7 +31,7 @@ const STYLES_COMMENT = css`
   white-space: nowrap;
   font-family: ${Constants.fontFamilies.mono};
   font-size: 13px;
-  color: ${colors.dark.gray[600]};
+  color: ${palette.dark.gray[600]};
   line-height: 150%;
 `;